Move getLocation to location service
